fix issue with upload image
