+++ index.js
@@ -8,7 +8,7 @@ const requestsRouter = require("./requests/requests-router");
 const PORT = process.env.PORT || 3001;
 const HOST = process.env.HOST || "localhost";
 const { verifyToken, validateUserToken } = require("./Middleware/auth");
-const profile = require("./storage/storage-upload");
+const profile = require("./storage/profile-img-upload-post");
 
 const app = express();
 
@@ -34,7 +34,7 @@ app.use("/users", verifyToken(), validateUserToken(), usersRouter);
 app.use("/users/rides", verifyToken(), validateUserToken(), usersRidesRouter);
 app.use("/rides", verifyToken(), validateUserToken(), ridesRouter);
 app.use("/rides/requests", verifyToken(), validateUserToken(), requestsRouter);
-app.use("/user/profile", profile);
+app.use("/profile", profile);
 
 app.use((err, req, res, next) => {
     console.log(err);
