+++ package.json
@@ -23,7 +23,7 @@
     },
     "homepage": "https://github.com/Lambda-School-Labs/carpal-be#readme",
     "dependencies": {
-        "aws-sdk": "^2.669.0",
+        "aws-sdk": "^2.670.0",
         "bcryptjs": "^2.4.3",
         "body-parser": "^1.19.0",
         "cors": "^2.8.5",
@@ -31,6 +31,7 @@
         "express": "^4.17.1",
         "express-fileupload": "^1.1.7-alpha.3",
         "helmet": "^3.22.0",
+        "jquery": "^3.5.1",
         "jsonwebtoken": "^8.5.1",
         "knex": "^0.21.1",
         "knex-cleaner": "^1.3.0",
